Nested Types: Generalize equi-join conjunct assignment for bushy plans.
In our equi-join predicate assignment logic, we used to assume that
the right-hand side or a join (or left-hand side if inverted) only contains
a single table ref id.

With the addition of subplans for nested types, we now sometimes generate
bushy plans where both sides of a join could cover multiple table ref ids.
This patch generalizes our existing logic to handle that case for equi-join
conjunct assignment, including inference of new join conjuncts as well
as removal of redundant conjuncts based on equivalence classes.

Change-Id: I252769ffd187988b1d617a17cd3c4ae638d5fcfa
Reviewed-on: http://gerrit.cloudera.org:8080/503
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins