Check for the existence of the psutil module before using it.
A previous change only enabled importing psutil if it was present on the system and always
returning True for check_process_exists. This broke the functionality of --kill wherein we
throw an error if the process has not been killed.

This patch adds a method to check for the existance of psutil and only checking for a
process being killed if it does.

Change-Id: I679ce12dc7e2732a8a95d5825c31d8a1bec354ec
Reviewed-on: http://gerrit.cloudera.org:8080/541
Reviewed-by: Ishaan Joshi <ishaan@cloudera.com>
Tested-by: Internal Jenkins