IMPALA-1389: only install SASL to /tmp if necessary
The SASL install directory was hardcoded to a path under /tmp. This solved
a very specific problem with ~'s in paths but could not be overridden. This
can lead to lost time when /tmp is cleared out periodically or on reboots.
IMPALA_CYRUS_SASL_INSTALL_DIR now defaults to a path under thirdparty unless
there is a tilde in the path. It can also be overridden by setting the
IMPALA_CYRUS_SASL_INSTALL_DIR environment variable.

Change-Id: I1aea2b51d265e3d1f04be0c915dcbee57c863be6
Reviewed-on: http://gerrit.cloudera.org:8080/536
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins