IMPALA-2188: expr-test fails on TestError() for built-ins/UDFs without Prepare() functions
When testing for an error, GetValue() in expr-test.cc never calls FetchResult() which is
responsible for ultimately calling the function corresponding to the builtin/UDF. It
only calls Exec() which only ultimately calls the Prepare() function responsible for that
builtin/UDF.

If a builtin without a Prepare() stage is tested for an error in expr-test, it will
always return with a Status::OK() because it never calls the actual function.
Changes have been made to accomodate testing of errors for such builtins/UDFs as well.

Change-Id: Id841bb5e03bc4f4255f56d64d05b890a6891bd4d
Reviewed-on: http://gerrit.cloudera.org:8080/612
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins