IMPALA-1983: Warn if table stats are potentially corrupt.
When the `numRows` parameter stored in the table properties is
errornously set to 0 and a number of non-empty files are present
the table statistics are considered to be corrupt.

To hint that there might be a problem, the explain statement will emit
an additional warning if it detects potentially corrupt table stats like
in the following example:

  Estimated Per-Host Requirements: Memory=42.00MB VCores=1
  WARNING: The following tables have potentially corrupt table and/or
  column statistics.
  compute_stats_db.corrupted

  03:AGGREGATE [FINALIZE]
  |  output: count:merge(*)
  |
  02:EXCHANGE [UNPARTITIONED]
  |
  01:AGGREGATE
  |  output: count(*)
  |
  00:SCAN HDFS [compute_stats_db.corrupted]
     partitions=1/2 files=1 size=24B

In addition, the small query optimization is disabled for such queries.

Change-Id: I0fa911f5132aa62195b854248663a94dcd8b14de
Reviewed-on: http://gerrit.cloudera.org:8080/689
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Tested-by: Internal Jenkins