Disable HDFS file handle caching by default
This patch modifies the Impala command line flag:

     --max_cached_file_handles=VAL

to disable caching of HDFS file handles if VAL is 0.

In addition, it moves the existing functional tests to a custom cluster
test and keeps a sanity check for no caching in the original
place. Furthermore, it will check that no file handles are leaked.

Change-Id: Ic36168bba52346674f57639e1ac216fd531b0fad
Reviewed-on: http://gerrit.cloudera.org:8080/691
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins