IMPALA-2251: disallow clashing text delimiters
It is currently possible to create tables with nonsensical delimited
text formats (e.g. same field and line delimiter). This patch adds
checks in the frontend to prevent creation of such text tables. It is
still possible for such tables to be created in Hive or for previously
created ones to be loaded from the Hive metastore.

Change-Id: Ia2979d1ecf8e880845b9f7394007f5e3ce2204c8
Reviewed-on: http://gerrit.cloudera.org:8080/694
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins