Add all replicas to tablet locations and use READ_LATEST
We're currently only using leader for reads, which leads to skew on scans
as they leaders, sometimes, are not evenly distributed across the cluster.

This makes computeScanRangeLocations() take all replicas into account, and
always pick the first one that is returned, whether it is the leader or not.
An exception is thrown if a tablet has no replicas.

This also changes the scanner's read mode to READ_LATEST. Currently we're
using SCAN_AT_SNAPSHOT, but we're not forwarding any timestamp, meaning each
replica is selecting its own, meaning the result is not a snapshot read at
all.

Since when scanning with READ_LATEST scans will not wait at all tests would
fail if we hit replicas, since the scans are ran right after the inserts. This
patch introduces a new flag that makes the scanner only scan leaders, for tests.

Change-Id: I5c2939a5c0d26a117df10d6464c7fed977ac9c46
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7907
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>