Checks for IMPALA-2314 and more accurate calculation of PHJ::LargestSpillPartition
This patch adds dchecks that ensure we push in hash_partitions_
allocated (ie non NULL) partition objects.

It also modifies the calculation of the largest partition at PHJ to
also account for the size of the probe stream. Previously, it was
only looking at the size of the build side.

Change-Id: I9fd9b7a89ac220682d41fbadcd67799169da3ab2
Reviewed-on: http://gerrit.cloudera.org:8080/790
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: Internal Jenkins