Gather and forward slot nullability from the frontend to the backend
When we create Kudu tables we set all non-key columns to nullable and
the key columns to be not nullable. However, when we scan the tables
we assume all slots are nullable. Ideally we'd allow nullability to
be set and stored in the metadata, but for now this patch makes it so
we also make columns/slots nullable on scans, exacly like we do when
we create tables.

This work will also make it slightly easier to allow to set nullability in
the metadata in the future.

Change-Id: Id1a507b75d03444b3ea3f0f634f5e7ffbb1a3eaa
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8009
Tested-by: jenkins
Reviewed-by: Martin Grund <mgrund@cloudera.com>