KUDU-916: add IGNORE to INSERT / UPDATE / DELETE
This adds a IGNORE flag in the SQL syntax to avoid failing on already
present or non-existing keys.

       INSERT IGNORE INTO ...
       UPDATE IGNORE SET x = y
       DELETE IGNORE FROM ...

For UPDATEs and DELETEs keys not found are ignored (e.g. when they were
concurrently deleted), for INSERTs duplicate keys are ignored. This
feature is mostly used in the context of concurrent, potentially
conflicting, write operations.

Change-Id: Ib7db05951b7ce2e94ad862857bfa8035e4b16ee4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8043
Tested-by: jenkins
Reviewed-by: Martin Grund <mgrund@cloudera.com>