IMPALA-2299: dedup zero-length tuples correctly
The dedup logic in row batch serialisation incorrectly assumed that two
distinct tuples must have two distinct memory addresses. This is not
true if one tuple has zero length.

Update the serialisation logic to check for this case and insert a
NULL.

Adds a unit test that exercises this bug prior to the fix and a query
test that also hit a DCHECK prior to the fix.

Change-Id: If163274b3a6c10f8ac6b6bc90eee9ec95830b7dd
Reviewed-on: http://gerrit.cloudera.org:8080/849
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Tested-by: Internal Jenkins