KUDU-930, KUDU-1106 - Support constant expressions on the right side of binary predicates
We had a bug that only when we would flip the binary predicate to normalize
it would we fold the constant expression into a literal. Meaning we would
only fold constant expressions on the left side of a BinaryPredicate and not
when the expressions were on the right side, causing us not to push those
predicates. This patch addresses this problem.

This also takes care of a long standing TODO by moving the normalization and
folding function into BinaryPredicate as a static public function.

Change-Id: I1e3b54001841646c3e06e7f390e410f7af328e30
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8104
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins