Flexible partitioning for Kudu tables.
This patch enables flexible partitioning for Kudu tables. The primary
goal of flexible partitioning is to specify dynamic partitioning
criteria so that as many tablets will be created as necessary to yield
the best performance.

In general there are two partitioning strategies HASH and RANGE
partitioning. Both can be used in conjunction or independent of each
other. The new syntax for the CREATE TABLE and CTAS statement are

       CREATE TABLE (...) DISTRIBUTE BY
       HASH(key_col1,key_col2) INTO n BUCKETS,
       HASH(key_col3) INTO m BUCKETS,
       RANGE(key_col1, key_col3) SPLIT ROWS (
         (1, 'a'), (2, 'b'), (3, 'c')
       );

The partitioning clause is started by the keywords DISTRIBUTE BY and
followed by zero or more HASH clauses followed by zero or more RANGE
clauses.

HASH clause:

A HASH clause may contain a list of referenced key columns. If no key
columns are specified all key columns are picked by default. Multiple
HASH clauses can be used the column list must not intersect and be a
subset of the key columns of the table. The second part of the HASH
clause defines into how many buckets the table is split. The bucket
count must be at least 2 or larger.

RANGE clause:

The RANGE clause defines a set of SPLIT ROWS that divide
a (sub-)key-range into multiple tablets. The column list of the RANGE
clause must reference a subset or all key columns of the table. A single
split row has as many components as referenced key columns and must be
of a compatible type. There can only be one RANGE clause per table.

Change-Id: Id5cab05add11fb206412f172d13fb9b35e04ad91
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7914
Tested-by: jenkins
Reviewed-by: Martin Grund <mgrund@cloudera.com>