KUDU-1080 - Keep Kudu scanners alive if the row batch queue is full
On a large self join query, we witnessed scanners on one of the sides
of the join blocking for a long time until the other side was completely
read. In this case the kudu scanners had been open and a little data had
been fetched but Impala went a long time without getting any more data.

The long period between consecutive NextBatch(), and subsequently NextBatch(),
calls would cause the scanners to time out on the tablet server, causing
"Scanner not found" error when the blocked side of the join was finally
allowed to proceed.

This patch addresses this by making the scanners ping the tablet server
if there is still data to fetch but GetNext() hasn't been called in a
while. We do so by introducing a new BlockingPutWithTimeout() method
to BlockingQueue() which we use when adding the batch to the queue.
If the put is successful we proceed as usual and get the next batch.
If it isn't we ping the tablet server to keep the scanner alive.

Change-Id: I6b6a0a2057008583982634a23c1b39bf8ad36352
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8245
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: David Alves <david.alves@cloudera.com>