IMPALA-2401: Properly set the number of nodes in UnnestNode and SingularRowSrcNode.
The bug was that we were hitting a Preconditions check inside
AnalyticPlanner.computeInputPartitionExprs() because the passed numNodes was -1,
but this function should expect a valid numNodes value. We passed a value of -1
as a result of not properly setting it in the computeStats() of UnnestNode
and SingularRowSrcNode.

We were setting the numNodes based on the numNodes of the containing SubplanNode
which is only initialized after the subplan tree (second child) has been set.
The fix is to set the numNodes based on the input of the containing SubplanNode.

Change-Id: Ib64ef75e7aaa08c4ea4c7e3c868189b35e5f1647
Reviewed-on: http://gerrit.cloudera.org:8080/906
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins