xfail test_verify_runtime_profile until IMPALA-2479 is fixed.
test_verify_runtime_profile in test_scanners.py expects the table
lineitem_multiblock to have 3 blocks. However, the snapshot image
does not load the table with the modified block size specified in
functional_schema_template.sql causing the test to fail in builds.

This patch xfails the test until we can figure out a way to
permanently fix the problem.

Change-Id: Ic646254cd28987646df9118195e1c29757ddbbda
Reviewed-on: http://gerrit.cloudera.org:8080/1152
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Tested-by: Internal Jenkins