IMPALA-2516: fix profile parser in hash join timer test
The time reported in the query profile for "ms" should be:
   123.456ms
The time parsing in the test was based on above format.

However, the intial patch of IMPALA-1800 trucated the time to:
   123ms
That caused the time parsing to fail.

In this fix, the time parsing is enhanced to handle this.

Change-Id: I4a38797e6c88d7101e53df5f69d9deab7c1f0819
Reviewed-on: http://gerrit.cloudera.org:8080/1218
Reviewed-by: Alan Choi <alan@cloudera.com>
Tested-by: Internal Jenkins