IMPALA-2535: PAGG hits mem_limit when switching to I/O buffers
The fix is lazy switching, similar to the PHJ fix in IMPALA-2480.
To make this work, we need a retry loop to handle the case where
appending to a spilled partition needs a switch to I/O buffers.

Testing: ran TPC-H Q18 scale factor 100 with 8000 mb memory limit.
Before patch it failed on most runs in the PAGG. After patch it always
succeds.

Change-Id: I30a047367cc918eaa6ca1dc74dfe82cc7406b19d
Reviewed-on: http://gerrit.cloudera.org:8080/1249
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins