IMPALA-2707: reduce hash table probes in agg
Add a FindBucket() method to the HashTable that returns an iterator
pointing to an empty bucket if the key is not in the hash table.
This avoids a second probe on Insert().

A simple benchmark showed a reduction of ~5% in time spent in the
aggregation in this query:

set num_nodes=1;
select l_orderkey, l_partkey, l_suppkey, sum(l_extendedprice)
from tpch_parquet.biglineitem
group by 1,2,3
order by 4 desc
limit 20;

Change-Id: I194f9e2c3acd604218cb34d19bc54681a682d478
Reviewed-on: http://gerrit.cloudera.org:8080/1555
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Readability: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins