IMPALA-2354: don't try to recover reserved block Pin error
The partitioned aggregation node previously attempted to recover from a
failed Pin of a reserved block by spilling. This scenario occurs only
when reservations in the block manager are overcommitted and as many
reservations have been filled as possible. This recovery is dangerous
because the block mgr may be left in an inconsistent state after
returning an error. Now we fail the query with MEM_LIMIT_EXCEEDED
in both the PartitionedAggregatedNode and PartitionedHashJoinNode.

Also remove a related TODO in PartitionedHashJoinNode.

Change-Id: I9ddd33ceb09ca3d41d38c8d5cbe6eafb24c583d4
Reviewed-on: http://gerrit.cloudera.org:8080/1554
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Readability: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins