IMPALA-2499: Evaluate a SELECT block using partition metadata.
This patch implements an optimization for evaluating table refs
with metadata instead of table scans for queries which satisfy
the following conditions:

(1) All scan slots being materialized are partition columns.
(2) The SELECT block only contains aggregate expressions bound
    by the these slots. The aggregate expressions should have
    distinct semantics such as SELECT MIN(X) from T; or
    SELECT COUNT(DISTINCT x) FROM T; If there are no aggregate
    expressions in the SELECT block, the query block must contain
    grouping expressions.

If the above conditions are satisfied, the scan nodes in the plan
of the SELECT block will be replaced with union nodes which materialize
the partition key values.

The query "select min(year), max(month) from functional.alltypes;"
went from 440ms on average to 10ms (44x speed up) with this change.
The speed-up depends on the number of rows per partition.

The following are plans before and after this change for the
following query:
select month, min(year) from functional.alltypes group by month

Before:

01:AGGREGATE [FINALIZE]
|  output: min(year)
|  group by: month
|
00:SCAN HDFS [functional.alltypes]
   partitions=24/24 files=24 size=478.45KB

After:

01:AGGREGATE [FINALIZE]
|  output: min(year)
|  group by: month
|
00:UNION
   constant-operands=24

This optimization is enabled by the query option 'optimize_partition_key_scans'.
Note that there are some caveats with this optimization. In particular,
the returned values may be inconsistent with that of the conventional
plans in the following two cases when this optimization is enabled.

1. If a user deletes a file without doing a refresh, the metadata
   becomes stale. The conventional plan will return an error in the
   scan (due to missing files). With this optimization, the partition
   key values of deleted partitions may be returned.

2. With the conventional plan, an empty partition will not be included
   in the evaluation of the return values. A partition is empty if it
   either contains (1) no file or (2) the file contains no row.
   This optimization may return different values in the second case
   above when there are no rows in the file.

Due to the potential inconsistencies above, users need to opt-in
for this optimization.

Change-Id: I30d4c7dab7610a30773fc60044499c468684dc9a
Reviewed-on: http://gerrit.cloudera.org:8080/1638
Reviewed-by: Michael Ho <kwho@cloudera.com>
Tested-by: Internal Jenkins