IMPALA-2701: Deprecate StringVal::Append()
StringVal::Append() is an error-prone interface. The intent
of StringVal is to be a place holder of (ptr, len). It's the
responsibility of the creator of the StringVal object to
manage the buffer pointed to by 'ptr'. In theory, a caller
can allocate a buffer, save the pointer to it and create a
StringVal with that pointer. If StringVal::Append() can free
or relocate the buffer without the caller's being aware of it,
the caller may end up writing to freed memory or double-free'ing
the buffer. To put it in another way, StringVal should not do
any memory management or it may cause subtle bugs.

This change deprecates StringVal::Append() from the public udf.h.
It was first introduced in 2.3. This change will break the backward
compatability of UDF SDK and it should be documented.

Change-Id: I0d5436884186d8b14ff97d30fbc395ef6b5962f3
Reviewed-on: http://gerrit.cloudera.org:8080/1578
Reviewed-by: Michael Ho <kwho@cloudera.com>
Tested-by: Internal Jenkins