IMPALA-2683: Do not link libstdc++ statically
Statically linking libstdc++ was an approach to reduce the number of
dynamically loaded dependencies. However, the linker would remove all
those symbols that are not referenced. In case that we would need a
symbol for runtime code generation that was previously removed, LLVM
would assert and fail.

This requires the presence of libgcc.so and libstdc++.so of the compiler
Impala was build with at runtime, which is the case for CDH packaging
and custom host builds.

Change-Id: I680478b95f60acc7085360be791bef9abdcd16eb
Reviewed-on: http://gerrit.cloudera.org:8080/1583
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins