Add EXTENDED explain plan to failed planner tests.
When planner tests fail in CI jobs, it is often difficult to diagnose
the issue because the STANDARD explain level used for the actual/expected
results lacks information about tuples/sizes/cardinality.

This patch adds the EXTENDED explain plan to the test error log for tests
that failed because if a mismatch in actual/expected results.

This patch also cleans up PlannerTest and and slightly changes its behavior
for better coverage. We now always compute both the single node and
distributed plans, regardless of whether those sections were specified
in the test. We only diff the actual/expected for existing sections, but
in any case we report exceptions that occurred during plan generation.

Change-Id: Id747251091c530b34b0f21079728851f0e17e4ce
Reviewed-on: http://gerrit.cloudera.org:8080/1694
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins