IMPALA-2750: Fix inline assembly for delimited text parser.
The change in eb9015e100 introduced a bug that when Impala is compiled
in release mode the inline assembly used to call the SSE 4.2 pcmpestrm
instruction would not work as advertised.

The reason is that the templated function yields different machine code
when compiled with O3 than the previous inlined function when multiple
calls to this function appear shortly after onther.

In the templated function GCC assumes that the result of the inlined
templated funciton is stored in xmm1 while it is in xmm0. When pointing
out to GCC that the result of the instruction call is in xmm0 using the
"Yz" constraint the final assembly is correct.

This patch adds an conditional compilation logic if the code is compiled
with LLVM or GCC to make sure the result is correctly stored.

Change-Id: Ie21cff7b3844a0bf9074fcafa2f32247b33bc2a5
Reviewed-on: http://gerrit.cloudera.org:8080/1691
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Tested-by: Internal Jenkins