IMPALA-2721: EE statestore tests: subscriber threads now solely handle transport close
In the statestore end-to-end tests, we create clients and servers to
simulate subscription to the statestore. The servers use threads to
handle the I/O. The input transport objects were shared resources whose
cleanup was being manipulated in both threads, leading to occasional
runtime errors.

In this case, let's just let the server threads handle cleanup. Doing it
in both is redundant.

Testing:

With repeated runs of the test_failure_detected test, I was able to
reproduce the bug within 1200 runs. After the fix, the test passed up to
8000 runs. Also, it's clear by inspection that the shared resources were
only touched in two places.

Change-Id: Ie878225f9715d0295ae43afaead958d95ad0e62a
Reviewed-on: http://gerrit.cloudera.org:8080/1841
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Internal Jenkins