IMPALA-2889: Compile only one binary for impalad, statestored, catalogd
This patch changes the build to produce only a single binary for
impalad, statestored and catalogd, and to use the name of the command
line process running the binary (i.e. 'argv[0]') to decide which daemon
to run.

By doing this, we can reduce the size of Impala's packages by nearly
two-thirds, which will reduce deployment times and Impala's
footprint (which is over 2GB in some cases).

Many of these benefits could be realised by dynamic linking our
libraries, but this change is relatively low risk as it doesn't change
the mechanism of dependency resolution, not does it change our cluster
management scripts. Dynamic linking can still provide us with a benefit
as a separate change.

Change-Id: Ic73dc906528beb69dc04a2e2fead0d8530d62e64
Reviewed-on: http://gerrit.cloudera.org:8080/2254
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Henry Robinson <henry@cloudera.com>