IMPALA-1755: Selectively enable codegen on scan nodes.
Codegen is not enabled in the scan node for certain formats
(e.g. parquet tables). In those cases, we may be stuck using
interpretation to evaluate the conjuncts bound by the scan
nodes due to lazy initialization of the codegen object.

This change modifies the planner to decide whether codegen
should always be enabled in scan nodes by estimating the cost
for evaluating the conjuncts on all rows in the table. The cost
is estimated by taking the product of number of rows and the
size of the conjuncts' expression trees. If the cost exceeds
certain threshold (determined by the query option
scan_node_codegen_threshold) or if stats is not available for
that table, codegen will be enabled for that scan node.

With this change, the following query with more than 500 items in
the in-predicate went from 6.5s to about 1.3s on average (5x speed up).

select l_partkey from tpch_parquet.lineitem
where substr(l_comment, 1, 5) in
('89436', '30868', '65085', '22977' ...)

Change-Id: Ic50a90f35416f4f833c9cbdc2d792c319486a74a
Reviewed-on: http://gerrit.cloudera.org:8080/2063
Reviewed-by: Michael Ho <kwho@cloudera.com>
Tested-by: Internal Jenkins