IMPALA-2721: HiveServer2 EE tests: session idle timeouts: account for wall clock time
The HiveServer2 end-to-end test for testing concurrent sessions with
mixed idle timeout values iterates a sequence of sessions with varying
timeout values, uses sleep() calls to force some sessions to become
idle, and then checks that queries against all sessions succeed or fail
based on whether the test expects that particular session to have timed
out.

Part of this iterating process involves a few operations to execute a
query statement via ImpalaHiveServer2Service.Client.ExecuteStatement()
or close a session via ImpalaHiveServer2Service.Client.CloseSession().
Unfortunately, the method by which the test decides whether a session is
expected to be active or idle does not take into account how long these
operations take. This means that in some corner cases--specifically code
coverage builds of Impala on CentOS--a session has idled out, correctly,
but the test doesn't expect that, so the test fails.

To get the test passing, the fix is to take wall clock time into account
by tracking the last time we got a response from each session. For a
given session, if the difference between the current time and the last
time we got a response is less than that session's idle timeout value,
we expect that session to still be active. If it's more, then we expect
that session to have gone idle.

Further investigation into why ExecuteStatement() takes notably long in
the code coverage/CentOS case is tracked in IMPALA-2951.

Also:
- Improve the name of the test to make its intention more clear.
- Remove all flake8 whitespace and useless import notices.

Testing:
Ran this using a private code coverage build in Jenkins, which uses
CentOS, and it passes. This includes a 300-iteration run just to make
sure. It still passes locally on Ubuntu.

Change-Id: If75127ede0f946e9f277f043a25b6d54bbeda554
Reviewed-on: http://gerrit.cloudera.org:8080/2058
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins