prefer Snappy from $IMPALA_TOOLCHAIN
Problem: impala-config.sh was setting the path to the Snappy library to
a path within thirdparty/. This assumption thus assumes thirdparty is
always being built. This is a bad assumption, because some build
sequences, like those to create nightly Docker images, rely on the
Impala toolchain and do not build thirdparty as part of their steps.
Moreover the whole point of the toolchain is to move away from needing
to build thirdparty.

Fix: Try to use Snappy from the $IMPALA_TOOLCHAIN first, and fall back
to Snappy from thirdparty otherwise. This is what's already done in
impala-config.sh with other libraries, like Thrift.

Testing:

1. Successful private Jenkins build passed

2. Obtained a Docker image where the following could produce the
problem:

./buildall.sh -notests -noclean -format -testdata

The data load would fail, because the HiveMetaStore process was not able
to locate the Snappy library. After applying the patch to the failing
Docker container, the entire command above succeeded.

3. Ensured all OSs' toolchains provided Snappy via manual download of
the Snappy package for each OS. I visually inspected that Snappy was
there and did not perform actual testing on all OSs. I only inspected
for the current $IMPALA_GCC_VERSION (4.9.2) and $IMPALA_SNAPPY_VERSION
(1.0.5).

Change-Id: I94445a52d98bce358a0acda01cd1bed4806db50c
Reviewed-on: http://gerrit.cloudera.org:8080/2226
Reviewed-by: Michael Brown <mikeb@cloudera.com>
Tested-by: Internal Jenkins