Leopard front end: protect cases where report and schedule paths are absent
If the leopard front end is run on a pristine system, the front end will
raise exceptions, because the directories containing reports and
schedules do not yet exist.

Since it's the controller's responsibility to create and populate these
filesystem artifacts, in the frontend, let's just protect against the
exception and log warnings.

Testing:

Before the fix, the reload_reports thread would immediately raise an
exception and the thread would die; loading the Web app in a browser
would cause an internal server error.

After the fix, the reload_reports thread logs a warning but continues to
run, and loading the Web app in a browser presents an appropriately
empty listing of run names and schedules.

Change-Id: I186174db136694d3df04d9159362c6eeaa70b5b8
Reviewed-on: http://gerrit.cloudera.org:8080/2164
Reviewed-by: Michael Brown <mikeb@cloudera.com>
Tested-by: Internal Jenkins