IMPALA-1822: Cleanup query profile log
Implemented log rotation for the query profile log, to prevent it
from growing unbounded, by adding log rotation to the SimpleLogger
class.

Previously, log rotation was only enabled for glog. It is still not
enabled for audit event logging or lineage logging, but this should
be simple to fix as they both use the same SimpleLogger class that
the profile logging uses.

There is a new flag, max_profile_log_files that controls how many
of the most recent profile log files are retained.

Change-Id: I52430d192d27cdf9bc0d81f26a63048eaeec5d29
Reviewed-on: http://gerrit.cloudera.org:8080/2260
Tested-by: Internal Jenkins
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>