IMPALA-3146: Initialise Coordinator::filter_mode_ in right place
The variable which records the filter mode for a coordinator was not
initialised before the first call to UpdateFilterRoutingTable() (which
was made for the coordinator fragment). This would occasionally disable
filters produced by the coordinator unexpectedly.

A test in the runtime filter test suite covers this bug, as it has a
coordinator fragment producing filters in GLOBAL mode.

Change-Id: I9cbad178d49e6ad5621b79565b10519d754c0b6c
Reviewed-on: http://gerrit.cloudera.org:8080/2466
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Reviewed-by: Silvius Rus <srus@cloudera.com>
Tested-by: Internal Jenkins