Reduce dependencies on inline header functions
This patch helps reduce compile times when modifying function
implementations in decimal-value.h, e.g. when tuning the implementations
of decimal operators. decimal-value.h was included in many places that
only need to know about the layout of DecimalValue, not the
implementation of decimal operations. It was included indirectly in many
files, e.g. via runtime-state.h.

The patch moves those functions to decimal-value.inline.h and is able to
avoid including decimal-value.inline.h in most headers. We also need to
do the same thing for raw-value.h and runtime-filter.h, because some
of the inline functions in raw-value.h referenced inline functions in
decimal-value.h, and functions in runtime-filter.h referenced inline
functions in runtime-filter.h.

It also moves timestamp parsing logic from .h to .cc file. This slightly
reduces the size of the llvm bitcode module and will slightly reduce
compile times. The functions are too large to benefit from inlining in
generated code.

Change-Id: Ic7a2f388cd14a4427c43af2724340a2ffe8fae3d
Reviewed-on: http://gerrit.cloudera.org:8080/2485
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins