IMPALA-3195: explicitly specify unique database in TestScanners.test_annotate_utf8_option
test_annotate_utf8_option was failing in the exhaustive build because
some other test was switching the session database from 'default' to
'functional_parquet'. This patch amends the patch to use the new
'unique_database' fixture, which both ensures the table will be
created in the expected database and handles table cleanup.

Change-Id: I981475581340edc0be68fa2813f5e63990202399
Reviewed-on: http://gerrit.cloudera.org:8080/2586
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: Skye Wanderman-Milne <skye@cloudera.com>